% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calculators.R
\name{alpha_adjust}
\alias{alpha_adjust}
\title{Compute alpha adjusted for multiplicity using Bonferroni or FDR placeholder}
\usage{
alpha_adjust(alpha, m = 1L, method = c("bonferroni", "none"))
}
\arguments{
\item{alpha}{Nominal alpha.}

\item{m}{Number of tests.}

\item{method}{One of "bonferroni" or "none". (FDR is typically used for discovery; for
confirmatory power, provide your familywise alpha or use bonferroni.)}
}
\value{
Adjusted alpha.
}
\description{
Compute alpha adjusted for multiplicity using Bonferroni or FDR placeholder
}
